practice_fastapi/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
pydantic
email-validator
pydantic[email]